requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.1.0
Werkzeug==0.15.4
